pyproject.toml
[tool.poetry]
name = "cresliant"
version = "1.1.0"
description = "A powerful node-based image editor made in Python."
authors = ["The Cresliant Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pillow = "^10.1.0"
dearpygui = "^1.10.1"
numpy = "^1.26.2"
pydantic = "^2.5.2"
psutil = "^5.9.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
pre-commit = "^3.6.0"
pyinstaller = "^6.3.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = ["__init__.py"]
extend-exclude = [".cache"]
fixable = ["ALL"]
select = ["ANN", "B", "C4", "DTZ", "E", "F", "I", "ISC", "INT", "PGH", "PIE", "Q", "RET", "RSE", "RUF", "S", "SIM", "TID", "UP", "W"]
ignore = [
    "ANN001", "ANN002", "ANN003", "ANN101", "ANN102", "ANN201", "ANN202", "ANN204", "ANN205", "ANN206", "ANN401",
    "B027", "B904",
    "C401", "C408",
    "D100", "D104", "D105", "D107", "D203", "D212", "D214", "D215", "D301",
    "D400", "D401", "D402", "D404", "D405", "D406", "D407", "D408", "D409", "D410", "D411", "D412", "D413", "D414", "D416", "D417",
    "E731",
    "ISC001",
    "RET503", "RET504",
    "RUF005", "RUF012", "RUF013", "RUF015",
    "S311", "S101", "S603",
    "SIM102", "SIM108"]
